[Add] `RendererType` and seperate `listTraitRenderer.ts`
