+++ src/base/browser/secondary/listWidget/listTrait.ts
@@ -1,8 +1,5 @@
-import { IListViewRenderer } from "src/base/browser/secondary/listView/listRenderer";
-import { ListItemType } from "src/base/browser/secondary/listView/listView";
 import { disposeAll, IDisposable } from "src/base/common/dispose";
 import { Emitter, Register } from "src/base/common/event";
-import { hash } from "src/base/common/hash";
 
 /**
  * The index changed in {@link ListTrait}.
@@ -102,35 +99,6 @@ export class ListTrait implements IDisposable {
     public dispose(): void {
         disposeAll([this._onDidChange]);
         this.indices.clear();
-    }
-
-}
-
-export class ListTraitRenderer<T> implements IListViewRenderer<T, HTMLElement> {
-
-    public readonly type: ListItemType;
-
-    private _trait: ListTrait;
-
-    constructor(trait: ListTrait) {
-        this._trait = trait;
-        this.type = hash(this._trait.trait);
-    }
-
-    public render(element: HTMLElement): HTMLElement {
-        return element;
-    }
-
-    public update(item: T, index: number, data: HTMLElement, size?: number): void {
-        if (this._trait.has(index)) {
-            data.classList.toggle(this._trait.trait, true);
-        } else {
-            data.classList.toggle(this._trait.trait, false);
-        }
-    }
-
-    public dispose(element: HTMLElement): void {
-        // do nothing
     }
 
 }
